#25 Improvements: 	- Add the location selection left menu in the map view navigation bar
